third_party/blink/web_tests/virtual/select-parser-relaxation/external/wpt/html/semantics/forms/the-select-element/customizable-select/select-option-hover-styles.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] Hover styles should be present for appearance:base-select options.
  assert_equals: appearance:base-select must be supported in order to run this test. expected "base-select" but got "auto"
Harness: the test ran to completion.
